# apps/hub_api/openapi.json
{
  "components": {
    "schemas": {
...
      "CacheDeletionResponse": {
        "properties": {
          "bytes_removed": {
            "title": "Bytes Removed",
            "type": "integer"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/CacheMutationDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          }
        },
        "required": [
          "processed",
          "bytes_removed",
          "details"
        ],
        "title": "CacheDeletionResponse",
        "type": "object"
      },
...
      "CacheMutationRequest": {
        "properties": {
          "entries": {
            "description": "Relative file paths within the HRRR cache.",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Entries",
            "type": "array"
          },
          "include_metadata": {
            "default": true,
            "description": "Also act on associated .grib2.json metadata files when touching GRIB assets.",
            "title": "Include Metadata",
            "type": "boolean"
          }
        },
        "required": [
          "entries"
        ],
        "title": "CacheMutationRequest",
        "type": "object"
      },
      "CacheStoreRequest": {
        "properties": {
          "entries": {
            "description": "Relative file paths within the HRRR cache.",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Entries",
            "type": "array"
          },
          "include_metadata": {
            "default": true,
            "description": "Also act on associated .grib2.json metadata files when touching GRIB assets.",
            "title": "Include Metadata",
            "type": "boolean"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional label appended to the archive folder name when storing files.",
            "title": "Label"
          }
        },
        "required": [
          "entries"
        ],
        "title": "CacheStoreRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "api__routes__weather__CacheEntriesResponse": {
        "properties": {
          "cache_dir": {
            "title": "Cache Dir",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/api__routes__weather__CacheEntryModel"
            },
            "title": "Entries",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "order": {
            "title": "Order",
            "type": "string"
          },
          "total_bytes": {
            "title": "Total Bytes",
            "type": "integer"
          },
          "total_files": {
            "title": "Total Files",
            "type": "integer"
          }
        },
        "required": [
          "cache_dir",
          "total_files",
          "total_bytes",
          "order",
          "limit",
          "entries"
        ],
        "title": "CacheEntriesResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ProjectPlant Hub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/weather/hrrr/cache": {
      "get": {
        "operationId": "inspect_hrrr_cache_api_v1_weather_hrrr_cache_get",
        "parameters": [
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "newest",
              "enum": [
                "newest",
                "oldest",
                "largest",
                "smallest"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__routes__weather__CacheEntriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect Hrrr Cache",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
    "/api/v1/weather/hrrr/cache/delete": {
      "post": {
        "operationId": "delete_hrrr_cache_entries_api_v1_weather_hrrr_cache_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheMutationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheDeletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Hrrr Cache Entries",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
    "/api/v1/weather/hrrr/cache/store": {
      "post": {
        "operationId": "store_hrrr_cache_entries_api_v1_weather_hrrr_cache_store_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheStoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheStoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Store Hrrr Cache Entries",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
    "/api/v1/weather/hrrr/fetch-log": {
      "get": {
        "operationId": "get_hrrr_fetch_log_api_v1_weather_hrrr_fetch_log_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HrrrFetchStatusModel"
                  },
                  "title": "Response Get Hrrr Fetch Log Api V1 Weather Hrrr Fetch Log Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hrrr Fetch Log",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
...
    "/api/v1/weather/hrrr/point": {
      "get": {
        "operationId": "get_hrrr_point_api_v1_weather_hrrr_point_get",
        "parameters": [
          {
            "description": "Force downloading the latest HRRR run for the point",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": true,
              "description": "Force downloading the latest HRRR run for the point",
              "title": "Refresh",
              "type": "boolean"
            }
          },
          {
            "description": "Persist the refreshed sample into the telemetry store",
            "in": "query",
            "name": "persist",
            "required": false,
            "schema": {
              "default": true,
              "description": "Persist the refreshed sample into the telemetry store",
              "title": "Persist",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90.0,
              "minimum": -90.0,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "maximum": 180.0,
              "minimum": -180.0,
              "title": "Lon",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrrrSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hrrr Point",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
    "/api/v1/weather/hrrr/refresh": {
      "post": {
        "operationId": "refresh_hrrr_point_api_v1_weather_hrrr_refresh_post",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90.0,
                  "minimum": -90.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180.0,
                  "minimum": -180.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lon"
            }
          },
          {
            "description": "Persist the refreshed sample into the telemetry store",
            "in": "query",
            "name": "persist",
            "required": false,
            "schema": {
              "default": true,
              "description": "Persist the refreshed sample into the telemetry store",
              "title": "Persist",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrrrSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Hrrr Point",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
    "/api/v1/weather/hrrr/schedule": {
      "post": {
        "operationId": "update_hrrr_schedule_api_v1_weather_hrrr_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HrrrScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrrrStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Hrrr Schedule",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
    "/api/v1/weather/hrrr/status": {
      "get": {
        "operationId": "get_hrrr_status_api_v1_weather_hrrr_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "history",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 200,
              "minimum": 1,
              "title": "History",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrrrStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Hrrr Status",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
    "/api/v1/weather/local": {
      "get": {
        "operationId": "get_local_weather_api_v1_weather_local_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90.0,
              "minimum": -90.0,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "maximum": 180.0,
              "minimum": -180.0,
              "title": "Lon",
              "type": "number"
            }
          },
          {
            "description": "Lookback window in hours",
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 6.0,
              "description": "Lookback window in hours",
              "title": "Hours",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Local Weather",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
    "/api/v1/weather/location": {
      "get": {
        "operationId": "get_active_weather_location_api_v1_weather_location_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherLocationResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Active Weather Location",
        "tags": [
          "v1",
          "weather"
        ]
      },
      "put": {
        "operationId": "upsert_active_weather_location_api_v1_weather_location_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherLocationUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherLocationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Active Weather Location",
        "tags": [
          "v1",
          "weather"
        ]
      }
    },
...
  }
}